Lambda-API/python-layer/requirements.txt
dnspython==2.7.0
numpy==2.2.6
pandas==2.2.3
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
tzdata==2025.2
